requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
django-filter==25.1
django-pytest==0.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==40.1.0
inflection==0.5.1
iniconfig==2.3.0
joblib==1.5.3
numpy==2.4.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.16.3
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
uritemplate==4.2.0
